requirements.txt
autopep8==1.4.4
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
WTForms==2.2.1
